+++ pom.xml
@@ -73,8 +73,21 @@
         <powermock.version>2.0.9</powermock.version>
         <selenium.version>4.17.0</selenium.version>
         <htmlunitdriver.version>4.17.0</htmlunitdriver.version>
+        <otel.version>1.39.0</otel.version>
         <copyright>Copyright (c) ${project.inceptionYear}-2024 ${project.organization.name}</copyright>
     </properties>
+
+    <dependencyManagement>
+        <dependencies>
+            <dependency>
+                <groupId>io.opentelemetry</groupId>
+                <artifactId>opentelemetry-bom</artifactId>
+                <version>${otel.version}</version>
+                <type>pom</type>
+                <scope>import</scope>
+            </dependency>
+        </dependencies>
+    </dependencyManagement>
 
     <build>
         <plugins>
@@ -347,6 +360,20 @@
             <groupId>org.seleniumhq.selenium</groupId>
             <artifactId>selenium-java</artifactId>
             <version>${selenium.version}</version>
+            <exclusions>
+                <exclusion>
+                    <groupId>io.opentelemetry</groupId>
+                    <artifactId>
+                        opentelemetry-sdk-extension-autoconfigure
+                    </artifactId>
+                </exclusion>
+                <exclusion>
+                    <groupId>io.opentelemetry</groupId>
+                    <artifactId>
+                        opentelemetry-sdk-extension-autoconfigure-spi
+                    </artifactId>
+                </exclusion>
+            </exclusions>
         </dependency>
         <dependency>
             <groupId>org.seleniumhq.selenium</groupId>
@@ -481,6 +508,14 @@
             <groupId>org.jline</groupId>
             <artifactId>jline</artifactId>
             <version>3.23.0</version>
+        </dependency>
+        <dependency>
+            <groupId>io.opentelemetry</groupId>
+            <artifactId>opentelemetry-sdk-logs</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>io.opentelemetry</groupId>
+            <artifactId>opentelemetry-exporter-otlp</artifactId>
         </dependency>
 
         <!-- ================= -->
