+++ src/main/java/com/xceptance/xlt/engine/metrics/MetricsReporter.java
@@ -1,0 +1,8 @@
+package com.xceptance.xlt.engine.metrics;
+
+import com.xceptance.xlt.api.engine.Data;
+
+public interface MetricsReporter
+{
+    public void reportMetrics(final Data data);
+}
